Add "timer go" to magnet_setup.msr
magnet_setup.msr initializes the magnet by stepping to different
isotopes. This physically warms up the magnet and gets it stabilized and
ready for measurements.

However, several scripts add a "timer halt" to wait for the warmup to
complete. An inattentive user(!) might miss the fact that the timer is
halted and waste time waiting for the script to continue.

Adding a "timer go" to the end of the magnet_setup.msr file solves this
problem.